Models: Port Mixins to own file
Split of mixins from app.user.models to their own file app.mixins
