+++ app/calendar/models.py
@@ -1,6 +1,6 @@
 from app import db
 from app.helpers import urlfriendly
-from app.user.models import LinkGenerator, SimpleAuditMixin
+from app.mixins import LinkGenerator, SimpleAuditMixin
 from flask import url_for
 
 class CalendarSetting(db.Model, SimpleAuditMixin):
